[Refactor] Package internal structure refactored:
1) Removed errors package;
2) Removed interfaces package;
3) Corrected UTs.

+++ internal/server/handler.go
@@ -1,22 +1,25 @@
-package grpc
+package server
 
 import (
 	"context"
+	"errors"
 
-	"github.com/omerkaya1/abf-guard/internal/domain/errors"
-	"github.com/omerkaya1/abf-guard/internal/grpc/api"
+	"github.com/omerkaya1/abf-guard/internal/db"
+	"github.com/omerkaya1/abf-guard/internal/server/api"
 )
+
+var errBadRequest = errors.New("bad request")
 
 // Authorisation is a handler for the GRPC Authorisation request
 func (s *ABFGuardServer) Authorisation(ctx context.Context, r *api.AuthorisationRequest) (*api.Response, error) {
 	if r == nil {
-		return nil, errors.ErrBadRequest
+		return nil, errBadRequest
 	}
 	err := s.Storage.GreenLightPass(ctx, r.GetIp())
 	switch err {
-	case errors.ErrDoesNotExist:
+	case db.ErrDoesNotExist:
 		return PrepareGRPCResponse(s.BucketManager.Dispatch(r.GetLogin(), r.GetPassword(), r.GetIp())), nil
-	case errors.ErrIsInTheBlacklist:
+	case db.ErrIsInTheBlacklist:
 		return PrepareGRPCResponse(false, nil), nil
 	case nil:
 		return PrepareGRPCResponse(true, nil), nil
@@ -28,18 +31,18 @@ func (s *ABFGuardServer) Authorisation(ctx context.Context, r *api.Authorisation
 // FlushBuckets is a handler for the GRPC FlushBuckets request
 func (s *ABFGuardServer) FlushBuckets(ctx context.Context, r *api.FlushBucketRequest) (*api.Response, error) {
 	if r == nil {
-		return nil, errors.ErrBadRequest
+		return nil, errBadRequest
 	}
 	if err := s.BucketManager.FlushBuckets(r.GetLogin(), r.GetIp()); err != nil {
-		return PrepareGRPCResponse(false, errors.ErrFlushBucketsFailed), nil
+		return PrepareGRPCResponse(false, err), nil
 	}
 	return PrepareGRPCResponse(true, nil), nil
 }
 
 // PurgeBucket is a handler for the GRPC PurgeBucket request
 func (s *ABFGuardServer) PurgeBucket(ctx context.Context, r *api.PurgeBucketRequest) (*api.Response, error) {
-	if r == nil {
-		return nil, errors.ErrBadRequest
+	if r == nil || r.GetName() == "" {
+		return nil, errBadRequest
 	}
 	if err := s.BucketManager.PurgeBucket(r.GetName()); err != nil {
 		return PrepareGRPCResponse(true, err), nil
@@ -49,8 +52,8 @@ func (s *ABFGuardServer) PurgeBucket(ctx context.Context, r *api.PurgeBucketRequ
 
 // AddIPToWhitelist is a handler for the GRPC AddIPToWhitelist request
 func (s *ABFGuardServer) AddIPToWhitelist(ctx context.Context, r *api.SubnetRequest) (*api.Response, error) {
-	if r == nil {
-		return nil, errors.ErrBadRequest
+	if r == nil || r.GetIp() == "" {
+		return nil, errBadRequest
 	}
 	if err := s.Storage.Add(ctx, r.GetIp(), r.GetList()); err != nil {
 		return PrepareGRPCResponse(false, nil), nil
@@ -60,8 +63,8 @@ func (s *ABFGuardServer) AddIPToWhitelist(ctx context.Context, r *api.SubnetRequ
 
 // DeleteIPFromWhitelist is a handler for the GRPC DeleteIPFromWhitelist request
 func (s *ABFGuardServer) DeleteIPFromWhitelist(ctx context.Context, r *api.SubnetRequest) (*api.Response, error) {
-	if r == nil {
-		return nil, errors.ErrBadRequest
+	if r == nil || r.GetIp() == "" {
+		return nil, errBadRequest
 	}
 	if err := s.Storage.Delete(ctx, r.GetIp(), r.GetList()); err != nil {
 		return PrepareGRPCResponse(false, nil), nil
@@ -71,8 +74,8 @@ func (s *ABFGuardServer) DeleteIPFromWhitelist(ctx context.Context, r *api.Subne
 
 // AddIPToBlacklist is a handler for the GRPC AddIPToBlacklist request
 func (s *ABFGuardServer) AddIPToBlacklist(ctx context.Context, r *api.SubnetRequest) (*api.Response, error) {
-	if r == nil {
-		return nil, errors.ErrBadRequest
+	if r == nil || r.GetIp() == "" {
+		return nil, errBadRequest
 	}
 	if err := s.Storage.Add(ctx, r.GetIp(), r.GetList()); err != nil {
 		return PrepareGRPCResponse(false, nil), nil
@@ -82,8 +85,8 @@ func (s *ABFGuardServer) AddIPToBlacklist(ctx context.Context, r *api.SubnetRequ
 
 // DeleteIPFromBlacklist is a handler for the GRPC DeleteIPFromBlacklist request
 func (s *ABFGuardServer) DeleteIPFromBlacklist(ctx context.Context, r *api.SubnetRequest) (*api.Response, error) {
-	if r == nil {
-		return nil, errors.ErrBadRequest
+	if r == nil || r.GetIp() == "" {
+		return nil, errBadRequest
 	}
 	if err := s.Storage.Delete(ctx, r.GetIp(), r.GetList()); err != nil {
 		return PrepareGRPCResponse(false, nil), nil
@@ -94,7 +97,7 @@ func (s *ABFGuardServer) DeleteIPFromBlacklist(ctx context.Context, r *api.Subne
 // GetIPList is a handler for the GRPC GetIPList request
 func (s *ABFGuardServer) GetIPList(ctx context.Context, r *api.ListRequest) (*api.ListResponse, error) {
 	if r == nil {
-		return nil, errors.ErrBadRequest
+		return nil, errBadRequest
 	}
 	return PrepareGRPCListIPResponse(s.Storage.GetIPList(ctx, r.GetListType()))
 }
